requirements.txt
llama-index
llama-index-llms-huggingface
llama-index-embeddings-huggingface
kss
evaluate
nltk
faststylometry
scikit-learn
jupyterlab
torch
transformers
faiss-cpu
numpy
accelerate
bitsandbytes
streamlit
pandas
plotly
streamlit-chat
streamlit-option-menu
streamlit-aggrid
konlpy
wordcloud
python-magic-bin
python-dateutil